Added: Voltage Areas + Electric Field Storage
Voltage Areas:

Every 2 seconds draws the voltage areas on the screen. Key: The lighter the color, the greater the charge. The darker the color, the lower the charge. Note: positive (+) voltage areas are in light colors such as pink, and negative(-) voltage areas are in dark colors such as black.
I assume there probably are some errors in calculation though I think its almost complete!

Note: I added a variable in the begining of the file:
double permitivity_of_free_space = 8.85418782 * Math.pow(10, -12);
Its used in the calculation for voltage.

Lastly, I changed the text near the balls to show their charge in micro-Columbs (µC) which is the standard unit of charge. By seeing their charge instead of acceleration, its easy to understand the voltage fields. I changed the small ball to have a + charge and the 2 big balls to a - charge. Look for: Around the small ball tons of light color rings (as its positive), and around the other balls for dark colors such as black (cuz they are negative). Last Note: The greater the magnitude of the charge (absolute value), the more affect it has on the voltage, and thus the greater the rings around it.