Fixed Acceleration Bug, BUT NOT QUITE
The Acceleration Problem:

They get insane acceleration when they are very close to each other. I think this is because in line 373, as part of the electro statics formula, we are dividing the force by the distance squared between the balls. If they are too close together, this results in a very low distance squared.

So: What I did was, make sure that distance squared has a min value of 100 (so even if they are close together, it doesn't go less that 100). This solves the problem - no more super speed balls - yet it is against the laws of physics, so we will have to ask Mr. Kuzsmaul to permit it.